Moved Address related values to separate dir
