+++ src/Value/Address/AddressId.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigipolisGent\Flanders\BasicRegisters\Value\Address;
+
+use DigipolisGent\Flanders\BasicRegisters\Value\AbstractId;
+
+/**
+ * Address object ID.
+ */
+final class AddressId extends AbstractId
+{
+
+}
