+++ src/Value/Address/Addresses.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace DigipolisGent\Flanders\BasicRegisters\Value;
+namespace DigipolisGent\Flanders\BasicRegisters\Value\Address;
 
 use DigipolisGent\Value\CollectionAbstract;
 
@@ -14,7 +14,7 @@ final class Addresses extends CollectionAbstract
     /**
      * Create a new collection.
      *
-     * @param \DigipolisGent\Flanders\BasicRegisters\Value\Address ...$addresses
+     * @param \DigipolisGent\Flanders\BasicRegisters\Value\Address\Address ...$addresses
      *   One or more addresses.
      */
     public function __construct(Address ...$addresses)
@@ -29,7 +29,7 @@ final class Addresses extends CollectionAbstract
     {
         $addresses = [];
 
-        /** @var \DigipolisGent\Flanders\BasicRegisters\Value\Address $address */
+        /** @var \DigipolisGent\Flanders\BasicRegisters\Value\Address\Address $address */
         foreach ($this->values as $address) {
             $addresses[] = (string) $address;
         }
