--- src/Value/AddressId.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace DigipolisGent\Flanders\BasicRegisters\Value;
-
-/**
- * Address object ID.
- */
-final class AddressId extends AbstractId
-{
-
-}
