+++ tests/Value/Address/AddressDetailTest.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace DigipolisGent\Tests\Flanders\BasicRegisters\Value;
+namespace DigipolisGent\Tests\Flanders\BasicRegisters\Value\Address;
 
-use DigipolisGent\Flanders\BasicRegisters\Value\Address;
-use DigipolisGent\Flanders\BasicRegisters\Value\AddressDetail;
-use DigipolisGent\Flanders\BasicRegisters\Value\AddressId;
-use DigipolisGent\Flanders\BasicRegisters\Value\FullAddress;
+use DigipolisGent\Flanders\BasicRegisters\Value\Address\Address;
+use DigipolisGent\Flanders\BasicRegisters\Value\Address\AddressDetail;
+use DigipolisGent\Flanders\BasicRegisters\Value\Address\AddressId;
+use DigipolisGent\Flanders\BasicRegisters\Value\Address\FullAddress;
 use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalName;
 use DigipolisGent\Flanders\BasicRegisters\Value\LanguageCode;
 use DigipolisGent\Flanders\BasicRegisters\Value\Locality\Locality;
@@ -21,7 +21,7 @@ use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameId;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers \DigipolisGent\Flanders\BasicRegisters\Value\AddressDetail
+ * @covers \DigipolisGent\Flanders\BasicRegisters\Value\Address\AddressDetail
  */
 class AddressDetailTest extends TestCase
 {
@@ -195,7 +195,7 @@ class AddressDetailTest extends TestCase
      * @param string $houseNumber
      * @param string $busNumber
      *
-     * @return \DigipolisGent\Flanders\BasicRegisters\Value\Address
+     * @return \DigipolisGent\Flanders\BasicRegisters\Value\Address\Address
      */
     private function createAddress(
         int $identitifier,
